Salad-Discord-Webhook/requirements.txt
python-dotenv==0.14.0
requests==2.24.0
win10toast==0.9
keyboard==0.13.5
pywin32==227
